Get request for api/zoos added
